split network code
